unpackage event name directories into commands
