+++ commands/index.js
@@ -1,0 +1,11 @@
+const events = [
+    require('./hi.js'),
+    require('./shutdown.js'),
+    require('./gameInviteEmbed'),
+    require('./profileEmbed.js'),
+    require('./help.js'),
+    require('./addMember.js'),
+    require('./removeMember.js'),
+]
+
+module.exports = events;
