--- commands/messageCreate/messageCreate_events.js
@@ -1,9 +1,0 @@
-const events = [
-    require('./hi.js'),
-    require('./shutdown.js'),
-    require('./gameInviteEmbed'),
-    require('./profileEmbed.js'),
-    require('./help.js'),
-]
-
-module.exports = events;
