--- commands/messageReactionAdd/messageReactionAdd_events.js
@@ -1,5 +1,0 @@
-const events = [
-    require('./addMember.js'),
-]
-
-module.exports = events;
