--- commands/messageReactionRemove/messageReactionRemove_events.js
@@ -1,7 +1,0 @@
-//messageReactionRemove
-
-const events = [
-    require('./removeMember.js'),
-]
-
-module.exports = events;
